# src/utils/swagger.yaml
swagger: '2.0'
info:
  description: 'Endpoint Swagger of RzBlog'
  version: '1.0.0'
  title: 'Swagger RzBlog 2.0'
  termsOfService: 'http://127.0.0.1:5000/api'
  contact:
    email: 'rzmobiledev@gmail.com'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: '127.0.0.1:5000'
basePath: '/api'
...
schemes:
  - 'http'
  - 'https'
securityDefinitions:
  api_key:
    type: 'apiKey'
    name: 'Authorization'
    in: 'header'
paths:
  /users:
    post:
      tags:
        - 'User'
      summary: 'Add a new user to database'
      description: 'User auth credential is required.'
      operationId: 'addUser'
      consumes:
        - 'application/json'
        - 'application/xml'
      produces:
        - 'application/json'
        - 'application/xml'
      parameters:
        - in: 'body'
          name: 'body'
          description: 'User object that needs to be added to the database'
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: 'Ok'
          schema:
            $ref: '#/definitions/User'
        '400':
          description: 'Unknown errors'
        '401':
          description: 'Email exists.'
      security:
        - api_key: []
    get:
      tags:
        - 'User'
      summary: 'List all users in database'
      description: 'Only superuser or admin can perform this action'
      operationId: 'listAllUser'
      consumes:
        - 'application/json'
        - 'application/xml'
      produces:
        - 'application/json'
        - 'application/xml'
      responses:
        '200':
          description: 'OK'
          schema:
            type: 'array'
            items:
              $ref: '#/definitions/User'
        '400':
          description: 'Unknown errors'
        '401':
          description: 'Unauthorized'
      security:
        - api_key: []
  /users/all:
    get:
      tags:
        - 'User'
      summary: 'List all active and softdeleted users in database'
      description: 'Only superuser or admin can perform this action'
      operationId: 'listSoftDeletedUser'
      consumes:
        - 'application/json'
        - 'application/xml'
      produces:
        - 'application/json'
        - 'application/xml'
      responses:
        '200':
          description: 'OK'
          schema:
            type: 'array'
            items:
              $ref: '#/definitions/User'
        '400':
          description: 'Unknown errors'
        '401':
          description: 'Access denied'
      security:
        - api_key: []
...
  /users/user/{id}:
    delete:
      tags:
        - 'User'
      summary: 'Hard Delete user'
      description: 'Logged in user has access to delete himself and superadmin or admin can delete all users. This operation will delete completely from database'
      operationId: 'hardDeleteUser'
      consumes:
        - 'application/json'
        - 'application/xml'
      produces:
        - 'application/xml'
        - 'application/json'
      parameters:
        - name: 'id'
          in: 'path'
          description: 'Initial user id'
          required: true
          type: 'integer'
      responses:
        '200':
          description: 'User deleted'
        '400':
          description: 'Unknown errors'
        '404':
          description: 'User not found'
        '403':
          description: 'Access Denied!'
      security:
        - api_key: []
...
definitions:
  User:
    type: 'object'
    properties:
      firstName:
        type: 'string'
        description: 'Your firstName:'
      lastName:
        type: 'string'
        description: 'Your lastName:'
      username:
        type: 'string'
        description: 'Your username'
      email:
        type: 'string'
        description: 'Email'
      password:
        type: 'string'
        format: 'password'
      role:
        type: 'integer'
        format: 'int64'
      active:
        type: 'boolean'
        default: false
    xml:
      name: 'User'
...